use flask blueprint
it enables the script runner to be imported inside any other
flask/django app and run inside a route.

another option would be to have this library to support multiple create
functions, e.g. `create_flask_app()` and `create_django_app()` and user
can call whichever one matches their use case?
